# dev-doc/swagger.yaml
basePath: /
definitions:
...
  chat.ChatMessageInfoResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      is_edited:
        type: boolean
      message_id:
        type: integer
      reply_to_id:
        type: integer
      sender_id:
        type: integer
      type:
        type: string
    type: object
  chat.ChatRoomMemberInfoResponse:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      joined_at:
        type: string
      last_read_at:
        type: string
      member_id:
        type: integer
      participant_id:
        type: integer
      role:
        type: string
    type: object
  chat.ChatRoomSummaryResponse:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      latest_message:
        type: string
      room_id:
        type: integer
      room_type:
        type: string
      unread_count:
        type: integer
    type: object
  chat.CreateRoomRequest:
    properties:
      allow_agent:
        type: boolean
      description:
        type: string
      max_members:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
  chat.CreateRoomResponse:
    properties:
      allow_agent:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      max_members:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
  chat.GetChatRoomDetailResponse:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      members:
        items:
          $ref: '#/definitions/chat.ChatRoomMemberInfoResponse'
        type: array
      messages:
        items:
          $ref: '#/definitions/chat.ChatMessageInfoResponse'
        type: array
      name:
        type: string
      room_id:
        type: integer
      room_type:
        type: string
    type: object
  chat.GetChatRoomMessagesResponse:
    properties:
      has_more:
        type: boolean
      messages:
        items:
          $ref: '#/definitions/chat.ChatMessageInfoResponse'
        type: array
    type: object
  chat.GetUserChatRoomsResponse:
    properties:
      bot:
        items:
          $ref: '#/definitions/chat.ChatRoomSummaryResponse'
        type: array
      channel:
        items:
          $ref: '#/definitions/chat.ChatRoomSummaryResponse'
        type: array
      direct:
        items:
          $ref: '#/definitions/chat.ChatRoomSummaryResponse'
        type: array
      group:
        items:
          $ref: '#/definitions/chat.ChatRoomSummaryResponse'
        type: array
    type: object
  chat.JoinRoomResponse:
    properties:
      invitation_id:
        type: integer
      joined_at:
        type: string
      member_id:
        type: integer
      role:
        type: string
      status:
        type: string
    type: object
  chat.MemberStatusInfoResponse:
    properties:
      last_read_at:
        type: string
      member_id:
        type: integer
    type: object
  chat.ResolveInvitationRequest:
    properties:
      approve:
        type: boolean
    type: object
  chat.ResolveInvitationResponse:
    properties:
      invitation_id:
        type: integer
      joined_at:
        type: string
      member_id:
        type: integer
      role:
        type: string
      status:
        type: string
    type: object
  chat.UpdateMemberStatusResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/chat.MemberStatusInfoResponse'
        type: array
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
host: dev.hiroliang.com
info:
  contact: {}
  description: Server for my all Goat application
  title: Goat-Server
  version: 1.0.0
paths:
...
  /api/chat/room:
    post:
      consumes:
      - application/json
      description: Create a new chat room (direct, group, channel, or bot). The creator
        becomes the owner member.
      parameters:
      - description: Room creation payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/chat.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chat.CreateRoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a chat room
      tags:
      - Chat
  /api/chat/room/{room_id}:
    get:
      description: Retrieve full detail of a chat room including member list and recent
        messages.
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.GetChatRoomDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Room Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get chat room detail
      tags:
      - Chat
  /api/chat/room/{room_id}/join:
    post:
      description: Join an existing chat room. For open rooms the member is created
        immediately; for invite-only rooms a pending invitation is created.
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.JoinRoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Room Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Already a member
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Join a chat room
      tags:
      - Chat
  /api/chat/room/{room_id}/member/status:
    patch:
      description: Mark the authenticated user's last-read timestamp for the given
        room, returning all members' statuses.
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.UpdateMemberStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Room Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update member read status
      tags:
      - Chat
  /api/chat/room/{room_id}/messages:
    get:
      description: Retrieve paginated messages for a chat room. Use before_id for
        cursor-based pagination.
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      - description: Return messages with ID less than this value (pagination cursor)
        in: query
        name: before_id
        type: integer
      - description: Maximum number of messages to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.GetChatRoomMessagesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Room Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get chat room messages
      tags:
      - Chat
  /api/chat/room/invitations/{invitation_id}:
    patch:
      consumes:
      - application/json
      description: Approve or reject a pending join invitation. Only the room owner/admin
        can approve.
      parameters:
      - description: Invitation ID
        in: path
        name: invitation_id
        required: true
        type: integer
      - description: Approval decision
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/chat.ResolveInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.ResolveInvitationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Invitation Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve a join invitation
      tags:
      - Chat
  /api/chat/rooms:
    get:
      description: Retrieve all chat rooms the authenticated user belongs to, grouped
        by room type (direct, group, channel, bot).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.GetUserChatRoomsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user chat rooms
      tags:
      - Chat
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"